reviewing docs, readme and module manifest
