+++ README.md
@@ -1,7 +1,7 @@
 <h1 align="center">PSNetScanners</h1>
 <div align="center">
-<sub>PowerShell ICMP and TCP async scanners</sub>
-<br /><br />
+  <sub>High-performance parallel ICMP and TCP scanners for PowerShell</sub>
+  <br /><br />
 
 [![build](https://github.com/santisq/PSNetScanners/actions/workflows/ci.yml/badge.svg)](https://github.com/santisq/PSNetScanners/actions/workflows/ci.yml)
 [![codecov](https://codecov.io/gh/santisq/PSNetScanners/branch/main/graph/badge.svg?token=b51IOhpLfQ)](https://codecov.io/gh/santisq/PSNetScanners)
@@ -10,17 +10,32 @@
 
 </div>
 
-PSNetScanners is a PowerShell Module that includes two cmdlets using async techniques for ICMP and TCP scanning. Essentially, like built-in `Test-Connection` and `Test-NetConnection` cmdlets with a few less options but much faster.
+**PSNetScanners** is a lightweight PowerShell module that provides fast, parallel network scanning capabilities using modern async .NET APIs.
+
+It includes two main cmdlets:
+
+- **`Test-PingAsync`** – Parallel ICMP echo requests (Ping)
+- **`Test-TcpAsync`** – Parallel TCP port scanning
+
+These cmdlets are designed as high-performance alternatives to `Test-Connection` and `Test-NetConnection -Port`, sacrificing some advanced options for significantly better speed when scanning multiple targets or ports.
+
+## Features
+
+- True parallel execution with configurable throttling
+- Async pattern using `Ping.SendPingAsync` and `TcpClient.ConnectAsync`
+- Never throws terminating errors (result-oriented design)
+- Excellent pipeline support (great for CSV input)
+- Compatible with Windows PowerShell 5.1 and PowerShell 7+
 
 ## Documentation
 
-Check out [__the docs__](./docs/en-US) for information about how to use this Module.
+Check out [**the docs**](./docs/en-US) for information about how to use this Module.
 
 ## Installation
 
 ### Gallery
 
-The module is available through the [PowerShell Gallery](https://www.powershellgallery.com/packages/PSNetScanners):
+This module is available through the [PowerShell Gallery](https://www.powershellgallery.com/packages/PSNetScanners):
 
 ```powershell
 Install-Module PSNetScanners -Scope CurrentUser
@@ -30,14 +45,27 @@ Install-Module PSNetScanners -Scope CurrentUser
 
 ```powershell
 git clone 'https://github.com/santisq/PSNetScanners.git'
-Set-Location ./PSNetScanners
+cd ./PSNetScanners
 ./build.ps1
 ```
 
 ## Requirements
 
-This module has no requirements and is fully compatible with __Windows PowerShell 5.1__ and [__PowerShell Core 7+__](https://github.com/PowerShell/PowerShell).
+- Windows PowerShell 5.1 or PowerShell 7+
+
+## Quick Examples
+
+```powershell
+# Fast ping sweep
+1..254 | ForEach-Object { "192.168.1.$_" } | Test-PingAsync -ThrottleLimit 100
+
+# Fast port scan
+Test-TcpAsync google.com 80,443,3389 -ConnectionTimeout 2000
+
+# Bulk scan from CSV
+Import-Csv .\targets.csv | Test-TcpAsync
+```
 
 ## Contributing
 
-Contributions are more than welcome, if you wish to contribute, fork this repository and submit a pull request with the changes.
+Contributions are welcome! Feel free to fork the repository and submit a Pull Request.
